requirements.txt
telethon==1.28.0 # Sesuaikan versi jika Anda menggunakan yang lain
aiohttp==3.9.5
python-dotenv==1.0.1
aiofiles==23.2.1
requests==2.31.0
beautifulsoup4==4.12.3